requirements.txt
gradio==3.19.1
numpy==1.22.0
transformers==4.26.1
wikipedia-api
torch 
torchvision 
torchaudio
openai
wikipedia
python-dotenv